+++ src/Components/Attendees.js
@@ -1,3 +1,33 @@
-export default function Attendees() {
-  return;
+import Attendee from './Attendee';
+import { useState } from 'react';
+export default function Attendees({people, attendees, event, updateEventAttendance}) {
+ // toggleEventAttendees, showAttendees
+  const [showAttendees, setShowAttendees] = useState(false);
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
+  
+  return (
+    <>
+      <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+
+      {showAttendees ? (
+        <div className="attendees">
+          {attendees.map((attendee, index) => (
+            <>
+              <Attendee 
+              attendee={attendee} 
+              index={index}
+              event = {event}
+              updateEventAttendance={updateEventAttendance}/>
+            </>
+          ))}
+        </div>
+      ) : null}
+      {/* // props ternary */}
+    </>
+  );
 }
